Move private headers to internal folder
