+++ include/khiva/internal/vectorUtil.h
