+++ src/khiva/matrix.cpp
@@ -4,9 +4,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+#include <khiva/internal/libraryInternal.h>
+#include <khiva/internal/matrixInternal.h>
 #include <khiva/matrix.h>
-#include "libraryInternal.h"
-#include "matrixInternal.h"
 
 #include <stdexcept>
